Deliverable3/RAG_DocQnA/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-mistralai
faiss-cpu
sentence-transformers
PyPDF2
